Fixed rebar.config to not use the git account for github.
When github allows public and anonymous access to code, use that
rather than relying on access through the 'git' account. Also,
many systems like the Mac OS X ship with broken or incomplete ssh
implementations; in this case the lack of /usr/libexec/ssh-askpass
which makes fetching dependencies a not so happy experience.