Handle gen_server:init return with timeout
It is common when initializing a gen_server to avoid potentially slow
setup in the init function itself and to use a short timeout to to
additional setup such as connecting to a server.

This patch makes wpool_process handle the case when gen_server:init
returns a tuple with a timeout.