handle custom strategy functions
This patch add the posisbility to pass a function to the strategy when
calling/casting/... a worker.

The function get the supervisor name as argument. To make it easy this patch
also expose the function find_wpool/1 to find the wpool record associated to a
name. Also wpool_pool:wpool_set/2, wpool_pool:wpool_get/2 have been added to
provide a transparent interface to the #wpool{} record.