Kill an overran worker
This feature allows to kill an overran task. As a proof of concept
the check condition on this commit is a maximum number of overrun
warning messages. The worker will be killed forcing the tasks
to be finished.

The original use case is to cancel a task after it had reached its
configured time out, avoiding the server to be loaded with tasks
which results will not be used by any client.