Fix wrong condition in tests
After running a task with `next_worker`, this task is run on the
currently indicated index in the `next` field, `1` in this test, and
then this field is atomically incremented (modulo the size), to `2` in
this test. So when asking for the new index, this _should_ be the
incremented value, two, so this test seems to have been covering an
inconsistency on how the storage is recovered.