Move from handmade cast_call to newer send_request
An important change is actually how the queue manager does the call: the
client would send a call to the queue manager with the whole request,
and the queue manager would then do a cast_call to the worker with the
whole request again, incurring in two copies of the request on each
erlang message send, which, for very big payloads, can mean lots of
copying. Now, the client, within the queue manager's API, calls the
queue manager for an available worker, and when it gets it, calls the
worker with the whole payload, doing the copies only once. This also
uncovers crashes that could happen for other strategies, that wouldn't
happen for the available_worker one as it was doing a very wide catch.