Implement wpool:send_request/ as in gen_server:send_request
Since OTP23, `gen_server` exposes some kind of "asynchronous calls",
that is, a mechanism to make a call, and do other stuff while we wait
for the result. `send_request` returns a reference, that can be used to
match any incoming answers later on, either using
`gen_server:wait_for_response/2`, or to do `gen_server:check_response/2`
for any message received in a `receive` statement or a `handle_info`
clause.