Optimise finding min_message_queue
If any process has a message queue of zero, we can already select that
one without having to explore all other processes, as lists:min would
have returned this anyway. Another optimisation is finding the minimum
in one pass over the lists: currently we map the queue lengths and then
we find the minimum, which is two passes over a list that can be
sufficiently long; we can find this in a single pass.