Migrate error_logger to logger
It has been a very long time since OTP made `logger` available and
deprecated `error_logger`, so it's about time we migrate to it.

Note too that we're choosing to report structured logging, that is,
maps, as `worker_pool` is a library, it should allow application code to
choose how to filter and format log messages.